+++ framework/aster-frame/src/arch/x86/mod.rs
@@ -8,7 +8,7 @@ pub mod iommu;
 pub(crate) mod irq;
 pub(crate) mod kernel;
 pub(crate) mod mm;
-pub(crate) mod pci;
+pub mod pci;
 pub mod qemu;
 #[cfg(feature = "intel_tdx")]
 pub(crate) mod tdx_guest;
