+++ framework/aster-frame/src/bus/mmio/device.rs
@@ -5,7 +5,7 @@ use log::info;
 
 use super::VIRTIO_MMIO_MAGIC;
 use crate::{
-    io_mem::IoMem,
+    device::io_mem::IoMem,
     trap::IrqLine,
     vm::{paddr_to_vaddr, Paddr, VmIo},
 };
