+++ framework/aster-frame/src/bus/mod.rs
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: MPL-2.0
 
 pub mod mmio;
-pub mod pci;
 
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub enum BusProbeError {
@@ -10,6 +9,5 @@ pub enum BusProbeError {
 }
 
 pub fn init() {
-    pci::init();
     mmio::init();
 }
