+++ kernel/comps/pci/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "aster-pci"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+aster-frame = { path = "../../../framework/aster-frame" }
+component = { path = "../../libs/comp-sys/component" }
+log = "0.4"
+spin = "0.9.4"
+bitflags = "1.3"
+align_ext = { path = "../../../framework/libs/align_ext" }
+
+[features]
