+++ kernel/comps/virtio/Cargo.toml
@@ -10,6 +10,7 @@ bitflags = "1.3"
 spin = "0.9.4"
 bytes = { version = "1.4.0", default-features = false }
 align_ext = { path = "../../../framework/libs/align_ext" }
+aster-pci = { path = "../pci" }
 aster-input = { path = "../input" }
 aster-block = { path = "../block" }
 aster-network = { path = "../network" }
