+++ kernel/comps/virtio/src/device/input/mod.rs
@@ -27,7 +27,7 @@
 //
 
 pub mod device;
-use aster_frame::io_mem::IoMem;
+use aster_frame::device::io_mem::IoMem;
 use aster_util::safe_ptr::SafePtr;
 use pod::Pod;
