+++ kernel/comps/virtio/src/transport/pci/common_cfg.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MPL-2.0
 
-use aster_frame::io_mem::IoMem;
+use aster_frame::device::io_mem::IoMem;
 use aster_util::safe_ptr::SafePtr;
 use pod::Pod;
