+++ kernel/comps/virtio/src/transport/pci/mod.rs
@@ -8,7 +8,7 @@ pub(super) mod msix;
 
 use alloc::sync::Arc;
 
-use aster_frame::bus::pci::PCI_BUS;
+use aster_pci::PCI_BUS;
 use spin::Once;
 
 use self::driver::VirtioPciDriver;
